+++ databricks-sdk-java/src/main/java/com/databricks/sdk/service/catalog/CatalogsAPI.java
@@ -83,6 +83,10 @@ public class CatalogsAPI {
    * specific ordering of the elements in the array.
    */
   public Iterable<CatalogInfo> list(ListCatalogsRequest request) {
+
+    if (request.getMaxResults() == null) {
+      request.setMaxResults(0L);
+    }
     return new Paginator<>(
         request,
         impl::list,