+++ databricks-sdk-java/src/main/java/com/databricks/sdk/service/catalog/ConnectionsAPI.java
@@ -85,6 +85,10 @@ public class ConnectionsAPI {
    * <p>List all connections.
    */
   public Iterable<ConnectionInfo> list(ListConnectionsRequest request) {
+
+    if (request.getMaxResults() == null) {
+      request.setMaxResults(0L);
+    }
     return new Paginator<>(
         request,
         impl::list,