+++ databricks-sdk-java/src/main/java/com/databricks/sdk/service/catalog/SchemasAPI.java
@@ -85,6 +85,10 @@ public class SchemasAPI {
    * the array.
    */
   public Iterable<SchemaInfo> list(ListSchemasRequest request) {
+
+    if (request.getMaxResults() == null) {
+      request.setMaxResults(0L);
+    }
     return new Paginator<>(
         request,
         impl::list,